+++ .gitignore
@@ -86,9 +86,3 @@ players/templates/players/game_ide.html
 
 # Game Frontend
 node_modules
-
-# Vagrant
-.vagrant/**
-
-# VSCode
-game_frontend/\.vscode/
