--- Pipfile
@@ -1,13 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[dev-packages]
-
-[packages]
-"e1839a8" = {path = ".", editable = true}
-kubernetes = "*"
-
-[requires]
-python_version = "2.7"
